# Do not apply conversation template during benchmarking